refactor(backend): inject dependencies on plan controller and service
