+++ backend/src/models/plan.ts
@@ -7,3 +7,16 @@ export interface Plan {
   city: string;
   dataCap: number; // franquia em GB
 }
+
+export const parseSpeed = (speed: string) => {
+  const unit = speed.replace(/\d/g, "");
+  const amountInMbps = parseInt(speed.replace(/\D/g, ""));
+
+  const multiplier =
+    {
+      Mbps: 1,
+      Gbps: 1000,
+    }[unit] ?? 1;
+
+  return multiplier * amountInMbps;
+};
