+++ backend/src/models/recommendedPlan.ts
@@ -1,0 +1,9 @@
+import { Plan } from "./plan";
+import { PlanPreferencesFields } from "./planPreferences";
+
+export type PreferencesMatches = Record<PlanPreferencesFields, boolean>;
+
+export interface RecommendedPlan extends Plan {
+  preferencesMatchRate: number;
+  preferencesMatches: PreferencesMatches;
+}
