+++ backend/src/test/mocks/app.ts
@@ -1,0 +1,19 @@
+import { createApp } from "../../app";
+import { createPlanController } from "../../controllers/planController";
+import { createInMemoryPlanService } from "../../services/inMemoryPlanService";
+
+export type App = ReturnType<typeof mockApp>;
+
+export const mockApp = () => {
+  const planService = createInMemoryPlanService();
+
+  const planController = createPlanController(planService);
+
+  return {
+    app: createApp(planController),
+
+    planService,
+
+    planController,
+  };
+};
